refactor code out of AbstractTask
